+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactoring: replaced custom list widget with native list view
 - Add Status, Message, and Peers columns to trackers table in torrent details
 - Reorder tracker columns: Tier, Host, Status, Peers, Seeders, Leechers, Downloads, Message
+- Add Port, Connection Type, and Direction columns to peers table in torrent details
 
 ### Removed